pyproject.toml
[tool.poetry]
name = "kairos-ontology-toolkit"
version = "1.1.3"
description = "Validation and projection tools for OWL/Turtle ontologies in the Kairos platform"
authors = ["Cnext <info@cnext.eu>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Cnext-eu/kairos-ontology-toolkit"
repository = "https://github.com/Cnext-eu/kairos-ontology-toolkit"
keywords = ["ontology", "owl", "turtle", "rdf", "shacl", "validation", "projection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "kairos_ontology", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
rdflib = "^7.4.0"
pySHACL = "^0.30.1"
Jinja2 = "^3.1.3"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0,<9.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.11"

[tool.poetry.group.service]
optional = true

[tool.poetry.group.service.dependencies]
fastapi = ">=0.115.0,<1.0.0"
uvicorn = {version = ">=0.30.0,<1.0.0", extras = ["standard"]}
httpx = ">=0.27.0,<1.0.0"
pydantic-settings = ">=2.0.0,<3.0.0"
sse-starlette = ">=2.0.0,<3.0.0"
github-copilot-sdk = ">=0.2.0,<1.0.0"
PyJWT = {version = ">=2.9.0,<3.0.0", extras = ["crypto"]}

[tool.poetry.scripts]
kairos-ontology = "kairos_ontology.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    # Ignore deprecation warnings from rdflib's SPARQL parser using old pyparsing API
    # This is a known issue in rdflib 7.5.0 with pyparsing 3.3.1
    # See: https://github.com/RDFLib/rdflib/issues/2563
    "ignore::DeprecationWarning:rdflib.plugins.sparql.parser",
    "ignore::DeprecationWarning:rdflib.plugins.sparql.parserutils",
    "ignore::DeprecationWarning:pyparsing.util",
]
